# Optimizations for VCL