--- a/src/autopas/containers/verletClusterLists/ClusterTower.h
+++ b/src/autopas/containers/verletClusterLists/ClusterTower.h
@@ -289,7 +289,13 @@ class ClusterTower : public ParticleCell<Particle> {
     _particleDeletionObserver = observer;
   };
 
- private:
+  /**
+   * Calls reserve on the underlying particle vector to allocate space for particles to come.
+   * @param n
+   */
+  void reserve(size_t n) { _particlesStorage.reserve(n); }
+
+  // private:
   /**
    * The number of particles in a full cluster.
    */