# Description

Make Region Iterators great (again?)!

- [x] Only search towers in Region of Interest
- [x] preallocate memory within towers upon generation

## Stats:

### Iterators
Scenario: `md-flexible --no-end-config --no-flops --container verletclusterlists --data-layout aos --newton3 enabled --particles-per-dimension 25 --traversal vcl_c06`

| Version | Threads  | Boundaries  | % of Simulation |
|---|---|---|---|
| old  | 1  | 19.644s  | 98.370% |
| old  | 8  | 19.847s  | 99.503% |
| new  | 1  |  0.150s | 28.162% |
| new  | 8  | 0.165s  | 50.928% |
| lc_c04_hcp | 1 | 0.040 | 11.095% |
| lc_c04_hcp | 8 | 0.048 | 34.968% |

Region Iterators are expected to be slower than for LC because instead of cells whole towers need to be searched.

### Preallocation

| Version | Threads  | rebuildNeighborLists [ns] |
|---|---|---|
| reserve 0  | 8  | 460477244 |
| reserve *2.7  | 8  |  384966463 |

## ~Related Pull Requests~

## Resolved Issues

- [x] fixes #438

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes.

- [x] Jenkins